# faster `mapreduce` for "reversed" views