Since we now have fast iteration over `FastContiguousSubArray` (#48720), it could make sense to implement a special method for `mapreduce` for "reversed" `FastSubArray`, that is, for those step with value -1 (`w2` below). Since for `mapreduce` we are allowed to process the array elements in any order, we can reverse the view to get something that has a much better chance of SIMD compiler optimization. Here are some benchmarks:
```
g(a) = reduce(+, a; init = zero(eltype(a)))
v = rand(Int8, 2^16);
w1 = @view v[begin:end];
w2 = @view v[end:-1:begin];

julia> @btime g($v); @btime g($w1); @btime g($w2)

# master
  574.189 ns (0 allocations: 0 bytes)
  641.550 ns (0 allocations: 0 bytes)
  834.255 μs (0 allocations: 0 bytes)
# this PR
  572.086 ns (0 allocations: 0 bytes)   # not affected by the PR
  675.769 ns (1 allocation: 48 bytes)   # around 30ns slower
  696.621 ns (1 allocation: 48 bytes)   # over 1000 times faster
```
The not implemented part of #48720 gave a speed-up factor of around 17 for `FastSubArray{Int8}`. Even if that is implemented later, this special method would still be around 70 times faster than that. Maybe this justifies to have a special method for this case. Maybe it doesn't.

**Note:** There is currently one issue with the PR. The new method allocates 48 bytes, which seems to take around 30ns (the time difference for `w1` above). I don't know how to get rid of this, but I'm hoping that somebody else has an idea. I have the feeling that it is somehow related to the treatment of keyword arguments.

Also note that I'm using `invoke` to avoid any interference with the implementation of `mapreduce` for `AbstractArray`.