@N5N3 2023-02-27
This is unacceptable for `op` without commutability. (e.g. `*` for `String`, `*` for `Matrix`, `vcat` for `Vector`)