@matthias314 2023-02-27
Good point! In principle this is easy to fix -- just replace `op` by `(x,y) -> op(y,x)`. Unfortunately, this seems to prevent the compiler from optimizing.