@mcabbott 2023-02-27
Or restrict to known commutative operations. IMO this should probably be hooked on later, to dispatch on `dims` rather than testing `if T == Colon`. Something similar is done for sums of Adjoint arrays, around here:

https://github.com/JuliaLang/julia/blob/8068e44190c297b7dde1fee4f947549865a9651c/stdlib/LinearAlgebra/src/adjtrans.jl#L384-L385