@matthias314 2023-02-27
> Or restrict to known commutative operations.

I agree. The whole point of this PR is to enable SIMD optimization for views with step value -1. For SIMD, I can only think of `BitInteger` types and commutative operations (`+`, `*`, `&`, `|`, maybe `min` and `max`).