@N5N3 2023-02-27
>  just replace op by (x,y) -> op(y,x).

We have `Base.FlipArgs`

> Unfortunately, this seems to prevent the compiler from optimizing.

This is a seperate problem. Once user provide `init`, `Base.reduce` always use `foldl` branch. (Thus disable many optimization)