@mikmoore 2023-02-27
> Or restrict to known commutative operations.

I really dislike this. It has no support or extensibility for new operations. It's not just single primitive operations on primitive types that can SIMD, it's also compositions of those operations. Likewise, it can support other types that wrap these primitive types for slightly different behavior (which may or may not change the behavior on the suggested operations). It's also unnecessarily fragile to evil users doing evil piracy, even if they've likely doomed themselves in other ways.

I'd much prefer we use a `Base.FlipArgs(op)` composition and let the compiler do its thing. If "the compiler doing its thing" is dumb, that's an orthogonal issue.