@matthias314 2023-03-02
Here is a new version. I got it working for possibly non-commutative operations. As far as I can tell, it was the implementation of `mapfoldr` that prevented the compiler from optimizing. I've added a (naive?) `mapfoldr` method for `AbstractArray`. Please check that I didn't miss any case to consider there. Also, I've slightly rearranged the code for `_mapreduce_dim` to reduce the number of methods needed to avoid method ambiguities.

Here are some benchmarks: For a given type `T`, I define
```
v = rand(T, 2^16); w = @view(v[end:-1:begin]); i = zero(T)
```
Then I get

| `T = Int8` | master | this PR |
|---|---|---|
| `foldl(+,w)` | 817.218 μs | 34.644 μs |
| `foldr(+,w)` | 812.553 μs | 654.147 ns |
| `reduce(+,w)` | 22.043 μs | 1.477 μs |
| `reduce(+,w; init=i)` | 818.110 μs | 659.043 ns |

| `T = Int64` | master | this PR |
|---|---|---|
| `foldl(+,w)` | 823.908 μs | 36.649 μs |
| `foldr(+,w)` | 833.575 μs | 7.908 μs |
| `reduce(+,w)` | 23.128 μs | 8.550 μs |
| `reduce(+,w; init=i)` | 813.772 μs | 7.922 μs |

| `T = Float64` | master | this PR |
|---|---|---|
| `foldl(+,w)` | 830.546 μs | 69.221 μs |
| `foldr(+,w)` | 835.472 μs | 7.941 μs |
| `reduce(+,w)` | 67.846 μs | 8.769 μs |
| `reduce(+,w; init=i)` | 839.744 μs | 7.951 μs |

For the last table I used `@fastmath`. I tested all `BitInteger` types with `+`, `*`, `&`, `|`, `xor`, `min`, `max` and `Float16`, `Float32` and `Float64` with `+`, `*`, `min`, `max`. The results are similar. I didn't see any allocations nor any significant change in runtime  compared to master when using the original vector `v` or `view(v, :)` instead of `w`. Well, in a few cases I got differences for `v` and `foldr` by a factor of up to 2, but I couldn't reproduce that.