@matthias314 2023-03-03
@N5N3 Do you think this is good now?

A related issue (outside the scope of this PR, I think) would be to understand why `reduce` without `init` is so much faster than `foldl` and `foldr`. Getting that speed for folding would dramatically improve performance for `FastSubArray` with `stride1` different from `1` and `-1`.