@DilumAluthge 2026-02-21
This PR has had no activity for two years+, so I'll close it as stale.

@matthias314 If you're still interested in working on this PR, let me know, and I can re-open it.